# stripe-api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stripe Integration API Documentation",
    "version": "1.0.0",
    "description": "API documentation for Stripe payment integration endpoints"
  },
  "paths": {
    "/api/stripe/create-checkout": {
      "post": {
        "summary": "Create a Stripe checkout session",
        "description": "Creates a Stripe checkout session for subscription purchases",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "SubscriptionPlanId": {
                    "type": "number",
                    "description": "ID of the subscription plan to purchase"
                  },
                  "PaymentMethod": {
                    "type": "string",
                    "enum": ["credit_card", "paypal", "bank_transfer", "crypto"],
                    "description": "Payment method to be used"
                  },
                  "AutoRenew": {
                    "type": "boolean",
                    "description": "Whether the subscription should auto-renew"
                  }
                },
                "required": ["SubscriptionPlanId"]
              },
              "example": {
                "SubscriptionPlanId": 1,
                "PaymentMethod": "credit_card",
                "AutoRenew": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": {
                      "type": "string",
                      "description": "Stripe checkout session ID"
                    },
                    "url": {
                      "type": "string",
                      "description": "URL to redirect user to Stripe checkout"
                    }
                  }
                },
                "example": {
                  "sessionId": "cs_test_a1b2c3d4e5f6g7h8i9j0",
                  "url": "https://checkout.stripe.com/pay/cs_test_a1b2c3d4e5f6g7h8i9j0"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  }
                },
                "example": {
                  "message": "Failed to create checkout session",
                  "statusCode": 400
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          }
        }
      }
    },
    "/api/stripe/webhook": {
      "post": {
        "summary": "Handle Stripe webhook events",
        "description": "Processes webhook events from Stripe (payment confirmations, failures, subscription updates)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Stripe webhook event payload"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "stripe-signature",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Stripe webhook signature for verification"
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook processed successfully",
            "content": {
              "application/json": {
                "example": {
                  "received": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid webhook payload",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid webhook signature",
                  "statusCode": 400
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    {
      "name": "stripe",
      "description": "Stripe payment integration endpoints"
    }
  ],
  "externalDocs": {
    "description": "Stripe API Documentation",
    "url": "https://stripe.com/docs/api"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ]
} 